Tags functionality and param refactoring
Add tags functionality to the module logic and refactor how path
parameters are handled in the module.

The initial implementation did not account for multiple path parameters;
meaning, we could not provide order of how path values should be used to
build API URL.

The tags functionality is WIP at the moment, but we want to add the
logic to public module so that local version of `PayProp::API::Public`
can be deprecated in favour of CPANM version.